feat(init): 自动初始化超级管理员账号。

+++ src/main/java/com/iam57/nyanime/pojo/entity/User.java
@@ -21,8 +21,8 @@ public class User implements Serializable {
     @Serial
     private static final long serialVersionUID = -4389108856096399423L;
     private Integer id;
-    private Byte sex;
-    private Byte userType;
+    private Integer sex;
+    private Integer userType;
     private String userName;
     private String password;
     private String avatar;
